+++ test/data_directory.dart
@@ -11,26 +11,23 @@
 ///   run things by default
 library data_directory;
 
-import "dart:io";
-import "package:path/path.dart" as path;
+import 'dart:io';
+
+import 'package:path/path.dart' as path;
 
 /// Returns whether [dir] is the root of the `intl` package. We validate that it
 /// is by looking for a pubspec file with the entry `name: intl`.
 bool _isIntlRoot(String dir) {
-  var file = new File(path.join(dir, 'pubspec.yaml'));
+  var file = File(path.join(dir, 'pubspec.yaml'));
   if (!file.existsSync()) return false;
   return file.readAsStringSync().contains('name: intl_translation\n');
 }
 
 String get packageDirectory {
-  // TODO(alanknight): We can also do this using mirrors, which may
-  // be better. See dart_style/test/utils.dart for an example.
-  var script = Platform.script;
-  // If being run in package:test, we're in an isolate with a data URI, so
-  // just assume the current directory is the package.
-  if (script.scheme == 'data') {
-    return Directory.current.path;
-  }
+  // Try the current directory.
+  if (_isIntlRoot(Directory.current.path)) return Directory.current.path;
+
+  // Search upwards from the script location.
   var dir = path.fromUri(Platform.script);
   var root = path.rootPrefix(dir);
 
@@ -38,6 +35,6 @@ String get packageDirectory {
     if (_isIntlRoot(dir)) return dir;
     dir = path.dirname(dir);
   }
-  throw new UnsupportedError(
+  throw UnsupportedError(
       'Cannot find the root directory of the `intl_translation` package.');
 }