+++ test/message_extraction/debug.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
+
 # The message_extraction_test.dart test uses a temporary directory and spawns
 # separate processes for each step. This can make it very painful to debug the
 # steps.
@@ -6,8 +7,12 @@
 # directory. You can run the script to run the test locally, or use this to
 # run individual steps or create them as launches in the editor.
 dart ../../bin/extract_to_arb.dart sample_with_messages.dart \
-part_of_sample_with_messages.dart
+  part_of_sample_with_messages.dart
+
 dart make_hardcoded_translation.dart intl_messages.arb
-dart ../../bin/generate_from_arb.dart --json --generated-file-prefix=foo_ \
-sample_with_messages.dart part_of_sample_with_messages.dart \
-translation_fr.arb french2.arb translation_de_DE.arb
+
+dart ../../bin/generate_from_arb.dart \
+  --json --generated-file-prefix=foo_ \
+  --null-safety \
+  sample_with_messages.dart part_of_sample_with_messages.dart \
+  translation_fr.arb french2.arb translation_de_DE.arb