+++ test/message_extraction/failed_extraction_test.dart
@@ -2,31 +2,33 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-@Timeout(const Duration(seconds: 180))
+@Timeout(Duration(seconds: 180))
 
 library failed_extraction_test;
 
-import "message_extraction_test.dart";
-import "dart:io";
-import "package:test/test.dart";
+import 'dart:io';
 
-main() {
-  test("Expect warnings but successful extraction", () {
-    runTestWithWarnings(warningsAreErrors: false, expectedExitCode: 0);
+import 'package:test/test.dart';
+
+import 'message_extraction_test.dart';
+
+void main() {
+  test('Expect warnings but successful extraction', () async {
+    await runTestWithWarnings(warningsAreErrors: false, expectedExitCode: 0);
   });
 }
 
-const List<String> defaultFiles = const [
-  "sample_with_messages.dart",
-  "part_of_sample_with_messages.dart"
+const List<String> defaultFiles = [
+  'sample_with_messages.dart',
+  'part_of_sample_with_messages.dart'
 ];
 
-void runTestWithWarnings(
-    {bool warningsAreErrors,
-    int expectedExitCode,
-    bool embeddedPlurals: true,
-    List<String> sourceFiles: defaultFiles}) {
-  verify(ProcessResult result) {
+Future<void> runTestWithWarnings(
+    {required bool warningsAreErrors,
+    int? expectedExitCode,
+    bool embeddedPlurals = true,
+    List<String> sourceFiles = defaultFiles}) async {
+  void verify(ProcessResult result) {
     try {
       expect(result.exitCode, expectedExitCode);
     } finally {
@@ -34,9 +36,10 @@ void runTestWithWarnings(
     }
   }
 
-  copyFilesToTempDirectory();
-  var program = asTestDirPath("../../bin/extract_to_arb.dart");
-  List<String> args = ["--output-dir=$tempDir"];
+  await copyFilesToTempDirectory();
+
+  var program = asTestDirPath('../../bin/extract_to_arb.dart');
+  var args = <String>['--output-dir=$tempDir'];
   if (warningsAreErrors) {
     args.add('--warnings-are-errors');
   }
@@ -44,10 +47,8 @@ void runTestWithWarnings(
     args.add('--no-embedded-plurals');
   }
   var files = sourceFiles.map(asTempDirPath).toList();
-  List<String> allArgs = [program]..addAll(args)..addAll(files);
+  var allArgs = <String?>[program, ...args, ...files];
   var callback = expectAsync1(verify);
 
   run(null, allArgs).then(callback);
 }
-
-typedef dynamic ThenArgument(ProcessResult _);