+++ CHANGELOG.md
@@ -1,5 +1,21 @@
+## 0.18.0
+  * Add support for Flutter locale split.
+  * Allow null safe code when parsing.
+  * Update analyzer dependency.
+  * Upgrade to `package:lints/recommended.yaml`.
+  * Initial null safety conversion.
+  * Remove petit_parser dependency.
+  * Address analyzer deprecations, see [#168](https://github.com/dart-lang/intl_translation/issues/168).
+  * Migrate to null safety.
+  
+## 0.17.10+1
+  * Generate code that passes analysis with `implicit-casts: false`.
+  * Allow use of `MessageExtraction` and `MessageGeneration` without `File`.
+  * Move arb generation from bin to lib so it's available to external packages.
+  * Update analyzer dependency.
+
 ## 0.17.10
-   * Update petitparser dependency.
+  * Update petitparser dependency.
 
 ## 0.17.9
   * Fix pub complaint trying to precompile a library file in bin by moving that file to lib/src.